1. Divided game logic into parts. 2. Another try to install grunt globally
